pyproject.toml
[project]
name = "CS2_Arb"
version = "0.1.0"
description = "CS2 skin arbitrage tool"
requires-python = ">=3.11"
authors = [{ name = "Jacob Huang" }]
license = { text = "Proprietary" }

dependencies = [
  "httpx==0.27.2",
  "python-dotenv==1.0.1",
  "tabulate==0.9.0",
  "pandas==2.2.3"
]

[tool.black]
line-length = 100
target-version = ["py311"]
skip-string-normalization = true

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I"]

[tool.setuptools]
packages = ["app"]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"